requirements.txt
pydantic>=2.0.0
fastapi>=0.95.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
httpx>=0.24.0
qdrant-client>=1.5.0
sentence-transformers>=2.2.2
deepseek-ai>=0.1.0
openai>=1.1.0
typing-extensions>=4.5.0
pyyaml>=6.0 